benchmark fix

+++ frame/contracts/src/benchmarking/mod.rs
@@ -420,7 +420,7 @@ benchmarks! {
 				},
 			],
 			call_body: Some(body::repeated_dyn(r * API_BENCHMARK_BATCH_SIZE, vec![
-				Counter(0, account_len), // address_ptr
+				Counter(0, account_len as u32), // address_ptr
 				Regular(Instruction::Call(0)),
 				Regular(Instruction::Drop),
 			])),
